UL-93: fix wording on HTML. it should be reworked
as it uses document.write() and i've seen that this is not recommended at
all. e.g. SO/4537963
It is most definitely "bad form." It only works with webpages if you
use it on the page load; and if you use it during runtime, it will re
place your entire document with the input. And if you're applying it
as strict XHTML structure it's not even valid code.

the alternatives:

    .innerHTML This is a wonderful alternative, but this attribute has
    to be attached to the element where you want to put the text.

    .createTextNode() is the alternative recommended by the W3C.